Create Classes Vs Instances
Write a Person class with an instance variable, and a constructor that takes an integer, as a parameter. The constructor must assign to after confirming the argument passed as is not negative; if a negative argument is passed as the constructor should set to and print Age is not valid, setting age to 0.. In addition, you must write the following instance methods:

yearPasses() should increase the instance variable by.
amIOld() should perform the following conditional actions:
If, print You are young.
If and, print You are a teenager.
Otherwise, print You are old.